Changed some of the Session Models to display hours and minutes
Modified MyRequest page and View Request page to now display that it is empty when it is empty

Fixed a grammar on _Session's Partial page

Added a new function whereby the program checks whether the session is over now. if it is over, it now allows user to mark the session as complete.

Marking as session complete function not implemented yet