Added a review system and reviews are reflected on users profile
changed the requirements of reviews to now display whether it belongs to a as a session owner or participant

Details page now displays the number of stars for the review as unicode instead of numbers (+ creative points)

After marking a session complete, it now checks whether there are participants, and if there are.. itll bring the session owner to give reviews to their participants. Afterwhich, a notificiation will be sent to the participants to to give a review to the session owner

Points are now only distributed if there are at least 1 participant in the session

Points are also distributed to the participants after they have attended the session

Added notification system for the user to see the notifications that they have. Currently it does not do anything when you click on "Do Action"

Removed the table header for profile when the person does not have any reviews

You can now distinctly tell which session is yours at the View All Sessions Page